functions/process-telegram-image/requirements.txt
azure-core==1.33.0
azure-functions==1.23.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
pycparser==2.22
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
